requirements.txt
aioredis==1.3.1
asgiref==3.2.7
astroid==2.2.5
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
autopep8==1.4.4
certifi==2019.6.16
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.9
daphne==2.4.1
defusedxml==0.6.0
Django==2.2.10
django-allauth==0.39.1
django-cors-headers==3.0.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
msgpack==0.6.2
oauthlib==3.0.1
pep8==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.3.1
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
zope.interface==5.1.0
